add refreshtoken endpoint

+++ src/utils/cloudinary.js
@@ -4,7 +4,7 @@ import fs from "fs"
 // file system whiuch helps to read, write or do anything with file
 
 cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  cloud_name:process.env.CLOUDINARY_CLOUD_NAME, 
   api_key:process.env.CLOUDINARY_API_KEY, 
   api_secret:process.env.CLOUDINARY_API_SECRET
 });
@@ -13,19 +13,25 @@ const uploadOnCloudinary= async (localFilePath )=>{
     try {
         if(!localFilePath) return null;
         // else upload file on cloudinary
-         const response= await cloudinary.v2.uploader.upload(localFilePath,
+         const response= await cloudinary.uploader.upload( localFilePath,
             {
                 resource_type: "auto"
                 // kon se type ki file hai ye khud dekh lena
             }
-        )
+        );
+          if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         // file uploaded successfully
         console.log("file uploaded on cloudinary",response.url);
+
         return response;
     }catch(error){
-      fs.unlinkSync(localFilePath)
+      if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
     //   remove the locally stored temporary file as the upload operation got failed
     return null;
     }
 }
-export {uploadOnCloudinary}
+export {uploadOnCloudinary};
